requirements.txt
mediapipe==0.10.14
numpy
opencv_python_headless==4.10.0.84
Pillow==10.4.0
streamlit==1.37.1
tensorflow==2.17.0
